Created InputSource class for keyboard and mouse functionality
Made an input class to create functionality for keyboard and mouse input.
Added getKey() checks if a key/button is being pressed or not
Added update() which updates the current keys/buttons being pressed and released.
In goat Engine.h a private variable of InputSource was added to it.
getInput() was added to GOAT_Engine.cpp to pull input information
pollEvents() was added to keep input updated inside of GOAT_Engine.cpp