Created Datastructures that will be used for the new renderng system
We are going to create a new design for our rendering. We are going to design the system so that sprites have values that are references to vertex values. The vertex values will be maintained as a vector so we don't need to constantly reconstruct a vector every draw call. With this optimization we should able to render ~30k independent sprites in one draw batch at 60 FPS